[WFCORE-4259] Prepare for the modularised WildFly Elytron.
