+++ elytron/src/main/java/org/wildfly/extension/elytron/AdvancedModifiableKeyStoreDecorator.java
@@ -82,12 +82,12 @@ import org.jboss.dmr.ModelType;
 import org.jboss.msc.service.ServiceController;
 import org.jboss.msc.service.ServiceName;
 import org.jboss.msc.service.ServiceRegistry;
+import org.wildfly.common.bytes.ByteStringBuilder;
 import org.wildfly.common.function.ExceptionSupplier;
 import org.wildfly.extension.elytron.FileAttributeDefinitions.PathResolver;
 import org.wildfly.extension.elytron._private.ElytronSubsystemMessages;
 import org.wildfly.security.credential.source.CredentialSource;
 import org.wildfly.security.pem.Pem;
-import org.wildfly.security.util.ByteStringBuilder;
 import org.wildfly.security.x500.X500;
 import org.wildfly.security.x500.cert.PKCS10CertificateSigningRequest;
 import org.wildfly.security.x500.cert.SelfSignedX509CertificateAndSigningKey;
